+++ Magic/src/main/java/com/elmakers/mine/bukkit/tasks/FinalizeImmediatelyTask.java
@@ -1,0 +1,16 @@
+package com.elmakers.mine.bukkit.tasks;
+
+import com.elmakers.mine.bukkit.magic.MagicController;
+
+public class FinalizeImmediatelyTask implements Runnable {
+    private final MagicController controller;
+
+    public FinalizeImmediatelyTask(MagicController controller) {
+        this.controller = controller;
+    }
+
+    @Override
+    public void run() {
+        controller.finalizeImmediately();
+    }
+}