+++ Magic/CHANGELOG.md
@@ -7,6 +7,7 @@
 
  - Add support for `disguise_target` to the `Disguise` action.
  - Add `take_wand: true` option to `TakeItem` action
+ - Add `mode: hidden` spell property to hide spells from the wand inventory
 
 # 10.8.1