+++ Magic/CHANGELOG.md
@@ -13,6 +13,7 @@
  - Add `take_wand: true` option to `TakeItem` action
  - Add `mode: hidden` spell property to hide spells from the wand inventory
  - Add custom `lore` spell option (does not currently support any templatization)
+ - Add `spells_use_wand_name: false` wand parameter
 
 # 10.8.1